Move contexts to own folder
