+++ app/components/SignatureCard/index.tsx
@@ -11,7 +11,7 @@ import GestureIcon from '@material-ui/icons/Gesture';
 import RemoveCircleIcon from '@material-ui/icons/RemoveCircle';
 
 import { UserType } from '@/common';
-import EditContext from '../EditContext';
+import EditContext from '../../context/EditContext';
 import { formatDate } from '../../utils/date';
 import SignatureType from './types';
