+++ app/views/ReportView/components/Appendices/index.tsx
@@ -6,7 +6,7 @@ import {
 
 import api, { ApiCallSet } from '../../../../services/api';
 import DataTable from '../../../../components/DataTable';
-import ReportContext from '../../../../components/ReportContext';
+import ReportContext from '../../../../context/ReportContext';
 import { AppendicesType, TcgaType, ComparatorType } from './types';
 import { sampleInformationColumnDefs, sequencingProtocolInformationColumnDefs, tcgaAcronymsColumnDefs } from './columnDefs';
 import GenomicReportOverview from './components/GenomicReportOverview';
