+++ app/views/ReportView/components/Discussion/components/CommentCard/__tests__/CommentCard.test.tsx
@@ -3,7 +3,7 @@ import {
   render, screen, fireEvent, cleanup, act,
 } from '@testing-library/react';
 
-import ReportContext from '@/components/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import SecurityContext from '@/components/SecurityContext';
 import { formatDate } from '@/utils/date';
 import CommentCard from '..';
