+++ app/views/ReportView/components/ProbeSummary/components/EventsEditDialog/index.tsx
@@ -10,8 +10,8 @@ import {
   Button,
 } from '@material-ui/core';
 
-import ReportContext from '../../../../../../components/ReportContext';
-import ConfirmContext from '../../../../../../components/ConfirmContext';
+import ReportContext from '../../../../../../context/ReportContext';
+import ConfirmContext from '../../../../../../context/ConfirmContext';
 import api from '../../../../../../services/api';
 
 type EventsEditDialogProps = {
