+++ app/views/ReportView/components/ProbeSummary/index.tsx
@@ -10,9 +10,9 @@ import EditIcon from '@material-ui/icons/Edit';
 
 import api, { ApiCallSet } from '@/services/api';
 import DataTable from '@/components/DataTable';
-import ReportContext, { ReportType, PatientInformationType, SampleInfoType } from '@/components/ReportContext';
-import EditContext from '@/components/EditContext';
-import ConfirmContext from '@/components/ConfirmContext';
+import ReportContext, { ReportType, PatientInformationType, SampleInfoType } from '@/context/ReportContext';
+import EditContext from '@/context/EditContext';
+import ConfirmContext from '@/context/ConfirmContext';
 import ReadOnlyTextField from '@/components/ReadOnlyTextField';
 import { formatDate } from '@/utils/date';
 import SignatureCard, { SignatureType } from '@/components/SignatureCard';
