+++ app/views/ReportView/components/Slides/components/UploadSlide/__tests__/UploadSlide.test.tsx
@@ -3,7 +3,7 @@ import {
   render, screen, fireEvent, cleanup, act,
 } from '@testing-library/react';
 
-import ReportContext from '@/components/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import UploadSlide from '..';
 
 const mockInputText = 'Sunday Television';
