+++ app/views/ReportView/components/Slides/components/UploadSlide/index.tsx
@@ -7,7 +7,7 @@ import {
 import { useSnackbar } from 'notistack';
 
 import api from '@/services/api';
-import ReportContext from '@/components/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import SlideType from '../../types';
 
 import './index.scss';
