+++ app/views/ReportView/components/TherapeuticTargets/index.tsx
@@ -7,10 +7,10 @@ import orderBy from 'lodash.orderby';
 import { useSnackbar } from 'notistack';
 
 import DataTable from '@/components/DataTable';
-import EditContext from '@/components/EditContext';
+import EditContext from '@/context/EditContext';
 import api from '@/services/api';
 import DemoDescription from '@/components/DemoDescription';
-import ReportContext from '@/components/ReportContext';
+import ReportContext from '@/context/ReportContext';
 import EditDialog from './components/EditDialog';
 import EvidenceHeader from './components/EvidenceHeader';
 import columnDefs from './columnDefs';
